partitioner/requirements.txt
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-tools==1.51.1
httplib2==0.21.0
idna==3.4
numpy==1.24.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.2
